refactor(nodebuilder/tests): move ipld test to ipld_test.go
